Move the Config classes for each DB table into its own file
